Fix compiler warnings and improve session cookie handling
- Fix SHA1 false positive warning by adding explicit length check
- Fix signed/unsigned comparison in check_post_header()
- Add CONFIG_LOG to LogChannels enum
- Add warning in setConfigInt for suspicious negative values (< -1)
- Remove FILE_TYPE_HTML restriction for session cookies
  (cookies now set for all response types when create_cookie is set)

+++ lib/src/system_sockets.c
@@ -266,7 +266,7 @@ static int check_post_header( socket_info* sock ){
 
 	//LOG(CONNECTION_LOG,NOTICE_LEVEL,sock->socket,"header->contentlenght %" PRIu64 " ",sock->header->contentlenght );
 
-	if ( sock->header->contentlenght > getConfigInt("max_post_size") ){
+	if ( sock->header->contentlenght > (uint64_t)getConfigInt("max_post_size") ){
 		LOG(CONNECTION_LOG,ERROR_LEVEL,sock->socket,"header->contentlenght > max_post_size ( %" PRIu64 " > %d ) ",sock->header->contentlenght, getConfigInt("max_post_size") );
 		return -1;
 	}
